+++ Editor/WithRuntime/GlobalEventSystem/GlobalEventManagerEditor.cs
@@ -1,0 +1,27 @@
+///스크립트 생성 일자 - 2025 - 03 - 12
+///스크립트 담당자 - 최현규
+///스크립트 생성 버전 - 0.1
+using UnityEditor;
+
+namespace CHG.EventDriven
+{
+	using Editor = UnityEditor.Editor;
+	
+	[CustomEditor(typeof(GlobalEventManager))]
+	public class GlobalEventManagerEditor : Editor
+	{
+		GlobalEventManager targetObject;		
+		
+		void OnEnable()
+		{
+		}
+		void OnDisable()
+		{
+		}
+		
+		public override void OnInspectorGUI()
+		{
+			DrawDefaultInspector();
+		}
+	}
+}