+++ Runtime/GlobalEventSystem/EventArgs/Base.cs
@@ -1,0 +1,15 @@
+///스크립트 생성 일자 - 2025 - 03 - 12
+///스크립트 담당자 - 최현규
+///스크립트 생성 버전 - 0.1
+
+using System;
+using UnityEngine;
+
+
+namespace CHG.EventDriven.Arguments
+{
+	public abstract class BaseEventArgs : EventArgs
+	{
+		
+	}
+}