+++ CHANGELOG.md
@@ -1,8 +1,13 @@
+# UtilityPackage - v0.0.9
+12 - 03 - 2025
+### 추가(Added)
+- 글로벌 이벤트 시스템, Event-Driven 구현
+
 # UtilityPackage - v0.0.8
 07 - 03 - 2025
 ### 추가(Added)
--충돌체 이벤트 핸들러Collider Event Handler
--절차적 이벤트 처리Sequential Event Scheduler
+- 충돌체 이벤트 핸들러Collider Event Handler
+- 절차적 이벤트 처리Sequential Event Scheduler
 
 ### 변경(Modified)
 ConditionalHide Attribute 개선